+++ core/src/main/java/io/grpc/util/SecretRoundRobinLoadBalancerProvider.java
@@ -18,7 +18,7 @@ package io.grpc.util;
 
 import io.grpc.LoadBalancer;
 import io.grpc.LoadBalancerProvider;
-import io.grpc.NameResolver.Helper.ConfigOrError;
+import io.grpc.NameResolver.ConfigOrError;
 import java.util.Map;
 
 /**
