+++ services/src/main/java/io/grpc/services/internal/HealthCheckingRoundRobinLoadBalancerProvider.java
@@ -22,7 +22,7 @@ import io.grpc.Internal;
 import io.grpc.LoadBalancer;
 import io.grpc.LoadBalancer.Helper;
 import io.grpc.LoadBalancerProvider;
-import io.grpc.NameResolver.Helper.ConfigOrError;
+import io.grpc.NameResolver.ConfigOrError;
 import io.grpc.services.HealthCheckingLoadBalancerUtil;
 import java.util.Map;
